+++ packages/csds-config/src/index.test.ts
@@ -1,4 +1,4 @@
-import * as index from "../src/index";
+import * as index from "./index";
 
 test("exposes DefaultCsdsConfig", () => {
   expect(index.DefaultCsdsConfig).toBeDefined();
