+++ packages/le-session-provider/tsconfig.build.json
@@ -5,6 +5,8 @@
     "outDir": "dist"
   },
   "exclude": [
-    "dist"
+    "dist",
+    "node_modules",
+    "**/*.test.ts"
   ]
 }
